Changed text conversion to allow for situation where a block elements are required

+++ test/unit/serverside/html2commonMark.spec.ts
@@ -14,16 +14,17 @@ function normalizeTree(root: commonmark.Node) {
 	let current: commonmark.WalkingStep;
 	while (current = walker.next()) {
 		let currentNode = current.node;
-		while (currentNode.type === 'Text' && currentNode.next && currentNode.next.type === 'Text') {
+		if (currentNode.type === 'Text' && currentNode.next && currentNode.next.type === 'Text') {
 			let newNode = new commonmark.Node('Text');
 			currentNode.parent.appendChild(newNode);
 			newNode.literal = currentNode.literal + currentNode.next.literal;
-			newNode.insertBefore(currentNode);
+			currentNode.insertBefore(newNode);
 			currentNode.next.unlink();
 			currentNode.unlink();
-			currentNode = newNode;
+			walker.resumeAt(newNode);
 		}
 	}
+	// console.log('normalized: ', new commonmark['XmlRenderer']().render(root));
 	return root;
 }
 
@@ -31,7 +32,6 @@ let assertEqual = (astExpected: commonmark.Node, astActual: commonmark.Node) => 
 	let expectedWalker = astExpected.walker();
 	let actualWalker = astActual.walker();
 	let expectedValue: commonmark.WalkingStep;
-
 
 	while (expectedValue = expectedWalker.next()) {
 		var actualValue = actualWalker.next();
@@ -48,8 +48,9 @@ let assertEqual = (astExpected: commonmark.Node, astActual: commonmark.Node) => 
 
 describe('CommonMark => html', () => {
 	// Var excluded: 7, 11
-	var scoped = [1, 2, 3, 4, 5, 6, 8, 9, 10, 12, 13];
-	// var scoped = [7];
+	var scoped = [1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 12, 13];
+	// var scoped = [11];
+	
 	tests.filter(t => scoped.indexOf(t.example) >= 0).forEach(test => {
 		it(`test #${test.example}, section ${test.section}: "${test.html }" ==> "${test.markdown}"`, (done) => {
 			sut.parse(test.html).then(result => {
